+++ packages/jivs-engine/src/Conditions/ConcreteConditions.ts
@@ -12,7 +12,7 @@
 import { LoggingCategory, LoggingLevel } from '../Interfaces/LoggerService';
 import { CodingError, assertNotNull } from '../Utilities/ErrorHandling';
 
-import type { IValueHost } from '../Interfaces/ValueHost';
+import { type IValueHost } from '../Interfaces/ValueHost';
 import { IValueHostsManager } from '../Interfaces/ValueHostsManager';
 import {
     type ICondition,
@@ -34,6 +34,7 @@ import { CompareToValueConditionBase, CompareToValueConditionBaseConfig } from '
 import { IValidatorsValueHostBase } from '../Interfaces/ValidatorsValueHostBase';
 import { toIInputValueHost } from '../ValueHosts/InputValueHost';
 import { NumberConditionBase, NumberConditionBaseConfig } from './NumberConditionBase';
+import { ConditionWithOneChildBase, ConditionWithOneChildBaseConfig } from './ConditionWithOneChildBase';
 
 
 /**
@@ -210,17 +211,6 @@ export interface RegExpConditionConfig extends RegExpConditionBaseConfig {
      * It is an alternative to expressionAsString. If both are supplied, this takes precedence.
      */
     expression?: RegExp;
-
-    /**
-     * Use the expression to find something wrong with the string instead of proving it to be valid.
-     * For example, if you want to tell the user "The period character is not allowed" and similar
-     * for invalid characters, you could have separate Validators using a regexp to detect the illegal
-     * character.
-     * When true, a string that matches the expression
-     * returns NoMatch instead of Match.
-     * When undefined, the value is false.
-     */
-    not?: boolean;    
 }
 
 /**
@@ -267,8 +257,6 @@ export class RegExpCondition extends RegExpConditionBase<RegExpConditionConfig>
     }
     protected evaluateString(text: string, valueHost: IValueHost, services: IValidationServices): ConditionEvaluateResult {
         let found = this.getRegExp(services).test(text);
-        if (this.config.not)
-            found = !found;
         return found ? ConditionEvaluateResult.Match : ConditionEvaluateResult.NoMatch;
     }    
 
@@ -913,4 +901,42 @@ export class MaxDecimalsCondition extends NumberConditionBase<MaxDecimalsConditi
     protected get defaultCategory(): ConditionCategory {
         return ConditionCategory.DataTypeCheck;
     }
+}
+
+/**
+ * ConditionConfig for {@link NotCondition}
+ */
+export interface NotConditionConfig extends ConditionWithOneChildBaseConfig
+{
+}
+
+/**
+ * Negates the result of a single child condition. Does nothing if the child condition
+ * results in Undetermined. Does not support Promises at this time.
+ */
+export class NotCondition extends ConditionWithOneChildBase<NotConditionConfig>
+{
+    public static get DefaultConditionType(): ConditionType { return ConditionType.Not; }
+
+    /**
+     * Evaluates the child condition and returns the opposite result, unless the child condition returns Undetermined.
+     * @param valueHost 
+     * @param valueHostsManager 
+     * @returns 
+     */
+    public evaluate(valueHost: IValueHost | null, valueHostsManager: IValueHostsManager): ConditionEvaluateResult | Promise<ConditionEvaluateResult> {
+        let condition = this.condition(valueHostsManager);
+        let result = condition.evaluate(valueHost, valueHostsManager);
+        if (result instanceof Promise)
+            throw new CodingError('Promises are not supported for child conditions at this time.');
+        switch (result) {
+            case ConditionEvaluateResult.Match:
+                return ConditionEvaluateResult.NoMatch;
+            case ConditionEvaluateResult.NoMatch:
+                return ConditionEvaluateResult.Match;
+            default:
+                return ConditionEvaluateResult.Undetermined;
+        }
+    }
+
 }