+++ packages/jivs-engine/tests/Conditions/Conditions.test.ts
@@ -1,13 +1,14 @@
+
 import { IDataTypeIdentifier } from './../../src/Interfaces/DataTypeIdentifier';
 import { ValueHostName } from "../../src/DataTypes/BasicTypes";
 import {
-    type RequireTextConditionConfig, type RangeConditionConfig, 
+    type RequireTextConditionConfig, type RangeConditionConfig,
     RequireTextCondition,
-    RangeCondition, EqualToCondition, 
+    RangeCondition, EqualToCondition,
     NotEqualToCondition, GreaterThanCondition,
-    GreaterThanOrEqualCondition,  LessThanCondition, 
-    LessThanOrEqualCondition, StringLengthConditionConfig,  StringLengthCondition,
-    RegExpConditionConfig, RegExpCondition, 
+    GreaterThanOrEqualCondition, LessThanCondition,
+    LessThanOrEqualCondition, StringLengthConditionConfig, StringLengthCondition,
+    RegExpConditionConfig, RegExpCondition,
     AllMatchCondition, DataTypeCheckConditionConfig, DataTypeCheckCondition,
     AnyMatchCondition, CountMatchesCondition,
     CountMatchesConditionConfig, AllMatchConditionConfig, AnyMatchConditionConfig, NotNullCondition, NotNullConditionConfig,
@@ -34,7 +35,9 @@ import {
     IntegerCondition,
     IntegerConditionConfig,
     MaxDecimalsCondition,
-    MaxDecimalsConditionConfig
+    MaxDecimalsConditionConfig,
+    NotCondition,
+    NotConditionConfig
 } from "../../src/Conditions/ConcreteConditions";
 
 import { LoggingCategory, LoggingLevel } from "../../src/Interfaces/LoggerService";
@@ -47,7 +50,7 @@ import { ConditionType } from "../../src/Conditions/ConditionTypes";
 import { LookupKey } from "../../src/DataTypes/LookupKeys";
 import { DataTypeConverterService } from "../../src/Services/DataTypeConverterService";
 import { IntegerConverter } from "../../src/DataTypes/DataTypeConverters";
-import { AlwaysMatchesConditionType, NeverMatchesConditionType, IsUndeterminedConditionType } from "../TestSupport/conditionsForTesting";
+import { AlwaysMatchesConditionType, NeverMatchesConditionType, IsUndeterminedConditionType, registerTestingOnlyConditions, EvaluatesAsPromiseConditionType } from "../TestSupport/conditionsForTesting";
 import { CompareToSecondValueHostConditionBaseConfig } from "../../src/Conditions/CompareToSecondValueHostConditionBase";
 import { CompareToValueConditionBaseConfig } from "../../src/Conditions/CompareToValueConditionBase";
 import { CapturingLogger } from "../TestSupport/CapturingLogger";
@@ -59,6 +62,7 @@ import { IValueHostsManager } from "../../src/Interfaces/ValueHostsManager";
 import { DataTypeIdentifierService } from '../../src/Services/DataTypeIdentifierService';
 import { IDataTypeConverter } from '../../src/Interfaces/DataTypeConverters';
 import { SimpleValueType } from '../../src/Interfaces/DataTypeConverterService';
+import { ConditionFactory } from '../../src/Conditions/ConditionFactory';
 
 
 describe('ConditionBase class additional cases', () => {
@@ -647,31 +651,6 @@ describe('class RegExpCondition', () => {
         vh.setValue('FirstLine\nABC\nLastLine');
         expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
     });
-    test('Set Config.not = true. evaluate returns Match if no match and NoMatch if matching', () => {
-        let services = new MockValidationServices(false, false);
-        let vm = new MockValidationManager(services);
-        let vh = vm.addMockInputValueHost(
-            'Property1', LookupKey.String, 'Label');
-        let config: RegExpConditionConfig = {
-            conditionType: ConditionType.RegExp,
-            valueHostName: 'Property1',
-            expressionAsString: 'ABC',
-            not: true
-        };
-        let testItem = new RegExpCondition(config);
-        vh.setValue('ABC');
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.NoMatch);
-        vh.setValue('ABCDEF');
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.NoMatch);
-        vh.setValue('zABC');
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.NoMatch);
-        vh.setValue(' ABC ');   // trim
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.NoMatch);
-        vh.setValue('abc'); // case sensitive failure
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
-        vh.setValue('AB\nC');
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
-    });
     test('evaluate not influenced by Config.trim=true', () => {
         let services = new MockValidationServices(false, false);
         let vm = new MockValidationManager(services);
@@ -700,27 +679,6 @@ describe('class RegExpCondition', () => {
             conditionType: ConditionType.RegExp,
             valueHostName: 'Property1',
             expressionAsString: 'ABC'
-        };
-        let testItem = new RegExpCondition(config);
-        vh.setInputValue(null);
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
-        vh.setInputValue(undefined);
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
-        vh.setInputValue(10);
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
-        vh.setInputValue(false);
-        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
-    });
-    test('With not=true, evaluate returns Undetermined for null, undefined, and non-string types', () => {
-        let services = new MockValidationServices(false, false);
-        let vm = new MockValidationManager(services);
-        let vh = vm.addMockInputValueHost(
-            'Property1', LookupKey.String, 'Label');
-        let config: RegExpConditionConfig = {
-            conditionType: ConditionType.RegExp,
-            valueHostName: 'Property1',
-            expressionAsString: 'ABC',
-            not: true
         };
         let testItem = new RegExpCondition(config);
         vh.setInputValue(null);
@@ -3878,6 +3836,38 @@ describe('class AllMatchCondition', () => {
         let testItem = new AllMatchCondition(config);
         expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
     });    
+    test('With 1 child that has invalid conditionConfig that evaluates as Undetermined and logs an error', () => {
+        let services = new MockValidationServices(true, true);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+        let config: AllMatchConditionConfig = {
+            conditionType: ConditionType.All,
+            conditionConfigs: [{
+                conditionType: 'Unknown'
+            }]
+        };
+        let testItem = new AllMatchCondition(config);
+
+        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
+        let logger = services.loggerService as CapturingLogger;
+        expect(logger.findMessage('Error creating condition', LoggingLevel.Error, null, null)).toBeTruthy();
+    });
+    test('With 1 child whose evaluate() function returns a Promise throws', () => {
+        let services = new MockValidationServices(true, true);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+        let config: AllMatchConditionConfig = {
+            conditionType: ConditionType.All,
+            conditionConfigs: [{
+                conditionType: EvaluatesAsPromiseConditionType
+            }]
+        };
+        let testItem = new AllMatchCondition(config);
+
+        expect(() => testItem.evaluate(vh, vm)).toThrow();
+    });
     test('category is Children', () => {
         let config: AllMatchConditionConfig = {
             conditionType: ConditionType.And,
@@ -4231,6 +4221,23 @@ describe('class AnyMatchCondition', () => {
         };
         let testItem = new AnyMatchCondition(config);
         expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
+    });
+    test('With 1 child that has invalid conditionConfig that evaluates as Undetermined and logs an error', () => {
+        let services = new MockValidationServices(true, true);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+        let config: AnyMatchConditionConfig = {
+            conditionType: ConditionType.Any,
+            conditionConfigs: [{
+                conditionType: 'Unknown'
+            }]
+        };
+        let testItem = new AnyMatchCondition(config);
+
+        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
+        let logger = services.loggerService as CapturingLogger;
+        expect(logger.findMessage('Error creating condition', LoggingLevel.Error, null, null)).toBeTruthy();
     });
     test('category is Children', () => {
         let config: AnyMatchConditionConfig = {
@@ -5013,4 +5020,168 @@ describe('MaxDecimalsCondition', () => {
         let testItem = new MaxDecimalsCondition(config);
         expect(testItem.category).toBe(ConditionCategory.Contents);
     });
+});
+
+describe('NotCondition', () => {
+    test('DefaultConditionType', () => {
+        expect(NotCondition.DefaultConditionType).toBe(ConditionType.Not);
+    });    
+    test('evaluate with valid child. Result of Match due to child NoMatch', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: NeverMatchesConditionType }
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(testItem.evaluate(null, vm)).toBe(ConditionEvaluateResult.Match);
+        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Match);
+    });
+    test('evaluate with valid child. Result of NoMatch due to child Match', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: AlwaysMatchesConditionType }
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(testItem.evaluate(null, vm)).toBe(ConditionEvaluateResult.NoMatch);
+        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.NoMatch);
+    });    
+    test('evaluate with valid child. Result of Undetermined due to child Undetermined', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: IsUndeterminedConditionType }
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(testItem.evaluate(null, vm)).toBe(ConditionEvaluateResult.Undetermined);
+        expect(testItem.evaluate(vh, vm)).toBe(ConditionEvaluateResult.Undetermined);
+    });
+    test('with invalid childconfig, logs error and evaluate returns undetermined', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: 'UnknownType' }
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(testItem.evaluate(null, vm)).toBe(ConditionEvaluateResult.Undetermined);
+        let logger = services.loggerService as CapturingLogger;
+        expect(logger.findMessage('Error creating condition', LoggingLevel.Error, null, null)).toBeTruthy();
+
+    });
+    test('with null childconfig, logs error and evaluate returns undetermined', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: null!
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(testItem.evaluate(null, vm)).toBe(ConditionEvaluateResult.Undetermined);
+        let logger = services.loggerService as CapturingLogger;
+        expect(logger.findMessage('childConditionConfig', LoggingLevel.Error, null, null)).toBeTruthy();
+
+    });    
+    test('with evaluate returning a promise in child condition, throws', () => {
+        let services = new MockValidationServices(false, true);
+        registerTestingOnlyConditions(services.conditionFactory as ConditionFactory);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: {
+                conditionType: EvaluatesAsPromiseConditionType
+            }
+        };
+
+        let testItem = new NotCondition(config);
+        
+        expect(() => testItem.evaluate(null, vm)).toThrow();
+
+    });    
+
+    test('category is Children', () => {
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: NeverMatchesConditionType }
+        };
+        let testItem = new NotCondition(config);
+        expect(testItem.category).toBe(ConditionCategory.Children);
+    });
+    test('category is overridden', () => {
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: NeverMatchesConditionType },
+            category: ConditionCategory.Contents
+        };
+        let testItem = new NotCondition(config);
+        expect(testItem.category).toBe(ConditionCategory.Contents);
+    });
+    test('gatherValueHostNames where child has a ValueHostName. Expect the one ValueHostName', () => {
+        let services = new MockValidationServices(true, true);
+        let vm = new MockValidationManager(services);
+
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: 
+                <RequireTextConditionConfig>{
+                    conditionType: ConditionType.RequireText,
+                    valueHostName: 'Field1'
+                }
+        };
+        let condition = new NotCondition(config);
+        let testItem = new Set<ValueHostName>();
+        expect(() => condition.gatherValueHostNames(testItem, vm)).not.toThrow();
+        expect(testItem.size).toBe(1);
+        expect(testItem.has('Field1')).toBe(true);
+    });        
+
+    test('dispose', () => {
+        let services = new MockValidationServices(false, false);
+        let vm = new MockValidationManager(services);
+        let vh = vm.addMockInputValueHost(
+            'Property1', LookupKey.String, 'Label');
+        let config: NotConditionConfig = {
+            conditionType: ConditionType.Not,
+            childConditionConfig: { conditionType: NeverMatchesConditionType }
+        };
+        let testItem = new NotCondition(config);
+        testItem.dispose();
+        expect(()=> testItem.evaluate(null, vm)).toThrow(TypeError);
+    });    
 });