+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   config were removed to let Modifier take over.
 - Extensive examples for configuration are now in jivs-examples.
 - **Breaking API change** Fluent syntax for all, any, and countMatches revised to take a function instead of a FluentConditionCollector
+- New condition, NotCondition. Inverts the result of a child condition. input('field').not((child)=> child.regExp(/\d*/));
 ## 0.17.1
 - Added setIssuesFound() method on ValidationManager and ValidatableValueHostBase. Simplifies how to send Jivs errors found by the server and sent up to the client.
 - Added {DataType} token to error messages, with localization using TextLocalizerService.